# Faster decoding of arrays in SpecificDatumReader